Modified indentation, and refactored some of the classes
Refactored BuildTriviaQuestions to TriviaQuestionsBuilder and GameStateManager to ScreenManager
