--- .idea/libraries/gdx_backend_robovm_1_9_4.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="gdx-backend-robovm-1.9.4">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.badlogicgames.gdx/gdx-backend-robovm/1.9.4/28822d4eac3c24b2b6d5ee50a017bc8981b7edd6/gdx-backend-robovm-1.9.4.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.badlogicgames.gdx/gdx-backend-robovm/1.9.4/efed340f61c78922b34d9525483709f425ecdd67/gdx-backend-robovm-1.9.4-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
