--- .idea/libraries/gdx_platform_1_9_4_natives_desktop.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="gdx-platform-1.9.4-natives-desktop">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.badlogicgames.gdx/gdx-platform/1.9.4/8e93afecc287ab152179983fc1b5f65db4c2fa4c/gdx-platform-1.9.4-natives-desktop.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
