--- .idea/libraries/gdx_platform_1_9_4_natives_ios.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="gdx-platform-1.9.4-natives-ios">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.badlogicgames.gdx/gdx-platform/1.9.4/934c60d5c1a1bd8fcf3583e3566e20b5a87184dd/gdx-platform-1.9.4-natives-ios.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
