--- .idea/libraries/jinput_platform_2_0_5_natives_linux.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="jinput-platform-2.0.5-natives-linux">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/net.java.jinput/jinput-platform/2.0.5/7ff832a6eb9ab6a767f1ade2b548092d0fa64795/jinput-platform-2.0.5-natives-linux.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
