--- .idea/libraries/jinput_platform_2_0_5_natives_osx.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="jinput-platform-2.0.5-natives-osx">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/net.java.jinput/jinput-platform/2.0.5/53f9c919f34d2ca9de8c51fc4e1e8282029a9232/jinput-platform-2.0.5-natives-osx.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
