--- .idea/libraries/jinput_platform_2_0_5_natives_windows.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="jinput-platform-2.0.5-natives-windows">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/net.java.jinput/jinput-platform/2.0.5/385ee093e01f587f30ee1c8a2ee7d408fd732e16/jinput-platform-2.0.5-natives-windows.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
