--- .idea/libraries/lwjgl_platform_2_9_2_natives_linux.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="lwjgl-platform-2.9.2-natives-linux">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.lwjgl.lwjgl/lwjgl-platform/2.9.2/d276cdf61fe2b516c7b7f4aa1b8dea91dbdc8d56/lwjgl-platform-2.9.2-natives-linux.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
