--- .idea/libraries/lwjgl_platform_2_9_2_natives_osx.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="lwjgl-platform-2.9.2-natives-osx">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.lwjgl.lwjgl/lwjgl-platform/2.9.2/d55b46b40b40249d627a83a7f7f22649709d70c3/lwjgl-platform-2.9.2-natives-osx.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
