--- .idea/libraries/lwjgl_platform_2_9_2_natives_windows.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="lwjgl-platform-2.9.2-natives-windows">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.lwjgl.lwjgl/lwjgl-platform/2.9.2/510c7d317f5e9e700b9cfaac5fd38bdebf0702e0/lwjgl-platform-2.9.2-natives-windows.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
