--- .idea/libraries/robovm_cocoatouch_2_2_0.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="robovm-cocoatouch-2.2.0">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-cocoatouch/2.2.0/ee379470f429ed37370c9896c4a066473ba2441e/robovm-cocoatouch-2.2.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-cocoatouch/2.2.0/52c2a821d8ff853a7daec5c4fe07ff77ffa321db/robovm-cocoatouch-2.2.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
