--- .idea/libraries/robovm_objc_2_2_0.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="robovm-objc-2.2.0">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-objc/2.2.0/68529e3192f2affd71d135acd2e9176f05d23c39/robovm-objc-2.2.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-objc/2.2.0/fb93a311e272f544822a7c873ffcf1143a2c7d88/robovm-objc-2.2.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
