--- .idea/libraries/robovm_rt_2_2_0.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="robovm-rt-2.2.0">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-rt/2.2.0/8a5cf08e2989ceaaa5d026c79cefbbc3cee56806/robovm-rt-2.2.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-rt/2.2.0/5f8f930aa5ec7b9a3d8eff3ad414ed6e0b8f01ff/robovm-rt-2.2.0-sources.jar!/" />
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-rt/2.2.0/5f8f930aa5ec7b9a3d8eff3ad414ed6e0b8f01ff/robovm-rt-2.2.0-sources.jar!/" />
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.mobidevelop.robovm/robovm-rt/2.2.0/5f8f930aa5ec7b9a3d8eff3ad414ed6e0b8f01ff/robovm-rt-2.2.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
