+++ core/src/main/java/GameObjects/TriviaQuestion.java
@@ -1,15 +1,15 @@
 package GameObjects;
 
 /**
- * Created by zach on 10/12/16.
- * TriviaQuestion holds information and errors related to each question that gets pulled in from the API.
+ * Created by zach on 10/12/16. TriviaQuestion holds information and errors related to each question
+ * that gets pulled in from the API.
  */
 
 public class TriviaQuestion {
 
-    public String question;
-    public String[] incorrectAnswers;
-    public String correctAnswer;
-    public String errorMessage;
+  public String question;
+  public String[] incorrectAnswers;
+  public String correctAnswer;
+  public String errorMessage;
 
 }
