+++ core/src/main/java/Scenes/Leaderboards.java
@@ -3,38 +3,31 @@ package Scenes;
 import com.badlogic.gdx.Screen;
 
 public class Leaderboards implements Screen {
-    @Override
-    public void show() {
+  @Override
+  public void show() {
+  }
 
-    }
+  @Override
+  public void render(float delta) {
+  }
 
-    @Override
-    public void render(float delta) {
+  @Override
+  public void resize(int width, int height) {
+  }
 
-    }
+  @Override
+  public void pause() {
+  }
 
-    @Override
-    public void resize(int width, int height) {
+  @Override
+  public void resume() {
+  }
 
-    }
+  @Override
+  public void hide() {
+  }
 
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
+  @Override
+  public void dispose() {
+  }
 }
