+++ core/src/main/java/Scenes/Lobby.java
@@ -2,39 +2,32 @@ package Scenes;
 
 import com.badlogic.gdx.Screen;
 
-public class Lobby implements Screen{
-    @Override
-    public void show() {
+public class Lobby implements Screen {
+  @Override
+  public void show() {
+  }
 
-    }
+  @Override
+  public void render(float delta) {
+  }
 
-    @Override
-    public void render(float delta) {
+  @Override
+  public void resize(int width, int height) {
+  }
 
-    }
+  @Override
+  public void pause() {
+  }
 
-    @Override
-    public void resize(int width, int height) {
+  @Override
+  public void resume() {
+  }
 
-    }
+  @Override
+  public void hide() {
+  }
 
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
+  @Override
+  public void dispose() {
+  }
 }
