+++ core/src/main/java/Scenes/MainMenu.java
@@ -2,43 +2,32 @@ package Scenes;
 
 import com.badlogic.gdx.Screen;
 
-/**
- * Created by zach on 10/6/16.
- */
-
 public class MainMenu implements Screen {
-    @Override
-    public void show() {
+  @Override
+  public void show() {
+  }
 
-    }
+  @Override
+  public void render(float delta) {
+  }
 
-    @Override
-    public void render(float delta) {
+  @Override
+  public void resize(int width, int height) {
+  }
 
-    }
+  @Override
+  public void pause() {
+  }
 
-    @Override
-    public void resize(int width, int height) {
+  @Override
+  public void resume() {
+  }
 
-    }
+  @Override
+  public void hide() {
+  }
 
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
+  @Override
+  public void dispose() {
+  }
 }
