+++ core/src/main/java/Scenes/TriviaQuestionsDisplay.java
@@ -3,43 +3,36 @@ package Scenes;
 import com.badlogic.gdx.Screen;
 
 /**
- * Created by zach on 10/13/16.
- * DisplayTriviaQuestions will build the GUI elements and display them for each question that is passed in.
+ * Created by zach on 10/13/16. DisplayTriviaQuestions will build the GUI elements and display them
+ * for each question that is passed in.
  */
 
 public class TriviaQuestionsDisplay implements Screen {
-    @Override
-    public void show() {
+  @Override
+  public void show() {
+  }
 
-    }
+  @Override
+  public void render(float delta) {
+  }
 
-    @Override
-    public void render(float delta) {
+  @Override
+  public void resize(int width, int height) {
+  }
 
-    }
+  @Override
+  public void pause() {
+  }
 
-    @Override
-    public void resize(int width, int height) {
+  @Override
+  public void resume() {
+  }
 
-    }
+  @Override
+  public void hide() {
+  }
 
-    @Override
-    public void pause() {
-
-    }
-
-    @Override
-    public void resume() {
-
-    }
-
-    @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void dispose() {
-
-    }
+  @Override
+  public void dispose() {
+  }
 }
