+++ core/src/test/java/Scenes/TestSceneTest.java
@@ -3,13 +3,13 @@ package Scenes;
 import org.junit.Test;
 
 /**
- * A test for the Test Scene class.
- * TODO(jmtaber129): Delete this test when the Test class is removed.
+ * A test for the Test Scene class. TODO(jmtaber129): Delete this test when the Test class is
+ * removed.
  */
 public class TestSceneTest {
-    @Test
-    public void show() throws Exception {
-        assert(true);
-    }
+  @Test
+  public void show() throws Exception {
+    assert (true);
+  }
 
 }
