+++ desktop/src/com/mygdx/game/desktop/DesktopLauncher.java
@@ -5,11 +5,11 @@ import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.mygdx.game.FilePlayMain;
 
 public class DesktopLauncher {
-	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.title = "File Play";
-		config.width = 480;
-		config.height = 800;
-		new LwjglApplication(new FilePlayMain(), config);
-	}
+  public static void main(String[] arg) {
+    LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+    config.title = "File Play";
+    config.width = 480;
+    config.height = 800;
+    new LwjglApplication(new FilePlayMain(), config);
+  }
 }
